requirements.txt
asn1crypto==1.5.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
filelock==3.14.0
idna==3.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pyarrow==16.1.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
requests==2.32.3
six==1.16.0
snowflake-connector-python==3.10.1
sortedcontainers==2.4.0
spotipy==2.24.0
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
